# optimization for txhashset rewind