--- a/chain/src/txhashset.rs
+++ b/chain/src/txhashset.rs
@@ -1234,8 +1234,12 @@ fn input_pos_to_rewind(
 				return None;
 			}
 		}
-		let bitmap =
-			Bitmap::fast_or(&block_input_bitmaps.iter().map(|x| x).collect::<Vec<&Bitmap>>());
+		let bitmap = Bitmap::fast_or(
+			&block_input_bitmaps
+				.iter()
+				.map(|x| x)
+				.collect::<Vec<&Bitmap>>(),
+		);
 		block_input_bitmaps.clear();
 		block_input_bitmaps.push(bitmap.clone());
 		Some(bitmap)