Optimization for `rewind()` function in `chain/src/txhashset.rs`. It's about as 700% faster as the old version, for this function call, in case the rewind block is very far from the Tip.

Compare between old version and new version:

- Test with block height 40,000:
```
# curl -0 -XGET  http://127.0.0.1:13413/v1/blocks/40000

Old Version:
Sep 14 20:09:05.106 DEBG txhashset: merkle_proof: output: Commitment(09d84a9c6b5939fbd12739421e6c99d26fc313f269046353d8eecf7879cbef1d56), block: 007453cb
Sep 14 20:09:06.070 DEBG merkle_proof  99312, last_pos 99312
                    <<<< 964ms spent

New Version:
Sep 14 20:31:54.413 DEBG txhashset: merkle_proof: output: Commitment(09d84a9c6b5939fbd12739421e6c99d26fc313f269046353d8eecf7879cbef1d56), block: 007453cb
Sep 14 20:31:54.548 DEBG merkle_proof  99312, last_pos 99312
                    <<<< 135ms spent
```
- Test with block height 50,000:
```
# curl -0 -XGET  http://127.0.0.1:13413/v1/blocks/50000

Old Version:
Sep 14 20:05:18.304 DEBG txhashset: merkle_proof: output: Commitment(09994dd2a6dd7cd7fbf9104e10c304cd8eaf16f74ae5d405e2b26c7e0e3b85c816), block: 02e79e8f
Sep 14 20:05:19.084 DEBG merkle_proof  119420, last_pos 119421
                    <<<<  780ms spent

New Version:
Sep 14 20:30:14.842 DEBG txhashset: merkle_proof: output: Commitment(09994dd2a6dd7cd7fbf9104e10c304cd8eaf16f74ae5d405e2b26c7e0e3b85c816), block: 02e79e8f
Sep 14 20:30:14.948 DEBG merkle_proof  119420, last_pos 119421
                    <<<< 106ms spent
```
- Test with block height 80,000:
```
# curl -0 -XGET  http://127.0.0.1:13413/v1/blocks/80000

Old Version:
Sep 14 20:04:21.345 DEBG txhashset: merkle_proof: output: Commitment(093941106eb2c3392e0f839dbd2d19926f71ae81d2368492f8b6b4b5854bb9bd79), block: 1316d80e
Sep 14 20:04:21.612 DEBG merkle_proof  185971, last_pos 185977
                    <<<< 267ms spent

New Version:
Sep 14 20:28:56.290 DEBG txhashset: merkle_proof: output: Commitment(093941106eb2c3392e0f839dbd2d19926f71ae81d2368492f8b6b4b5854bb9bd79), block: 1316d80e
Sep 14 20:28:56.325 DEBG merkle_proof  185971, last_pos 185977
                    <<<< 35ms spent
```

**another** optimization is for `Bitmap`, use `Bitmap::fast_or()` to replace `or_inplace()`. This was a `TODO` item in the old code. I also add a bench test for it:
```
$ cd store
$ cargo test --release bench_fast_or -- --nocapture

running 1 test
  or_inplace():     5.098ms. bitmap cardinality: 225785
     fast_or():     0.778ms. bitmap cardinality: 226018
fast_or_heap():     9.203ms. bitmap cardinality: 225628
```

A main optimization is using `get_hash_by_height()` instead of `get_block_header()`.

BTW, this PR came from the analysis result of issue: https://github.com/mimblewimble/grin/issues/1477