@garyyu 2018-09-14
updated for the review comments:
- add code comments for function `bitmap_fast_or`
- remove `map(|x| x)`